[rl] Move OpenReward under integrations
