+++ lib/marin/src/marin/rl/integrations/openreward/__init__.py
@@ -1,27 +1,35 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-from .task_manifest import (
-    OpenRewardPromptBlock,
-    OpenRewardPromptBlockType,
-    OpenRewardTaskManifest,
-    OpenRewardTaskManifestEntry,
-    OpenRewardToolSpec,
-    SecretsMapping,
+from marin.rl.integrations.openreward.client import load_openreward_client
+from marin.rl.integrations.openreward.env import OpenRewardEnv
+from marin.rl.integrations.openreward.manifest import (
     build_openreward_task_manifest,
-    load_openreward_client,
     load_openreward_task_manifest,
     prepare_openreward_task_manifest,
     resolve_task_indices,
     save_openreward_task_manifest,
 )
+from marin.rl.integrations.openreward.models import (
+    JSONDict,
+    OpenRewardPromptBlock,
+    OpenRewardPromptBlockType,
+    OpenRewardTaskManifest,
+    OpenRewardTaskManifestEntry,
+    OpenRewardToolSpec,
+    SecretValue,
+    SecretsMapping,
+)
 
 __all__ = [
+    "JSONDict",
+    "OpenRewardEnv",
     "OpenRewardPromptBlock",
     "OpenRewardPromptBlockType",
     "OpenRewardTaskManifest",
     "OpenRewardTaskManifestEntry",
     "OpenRewardToolSpec",
+    "SecretValue",
     "SecretsMapping",
     "build_openreward_task_manifest",
     "load_openreward_client",
