+++ lib/marin/src/marin/rl/integrations/openreward/manifest.py
@@ -3,82 +3,19 @@
 
 import json
 from collections.abc import Mapping, Sequence
-from enum import StrEnum
 from typing import Any, Protocol
-
-from pydantic import BaseModel, ConfigDict, model_validator
 
 from rigging.filesystem import open_url
 
-JSONDict = dict[str, Any]
-SecretValue = str | tuple[str, list[str]]
-SecretsMapping = Mapping[str, SecretValue]
-
-
-class _StrictModel(BaseModel):
-    """Immutable pydantic base with strict field handling."""
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-
-class OpenRewardPromptBlockType(StrEnum):
-    TEXT = "text"
-    IMAGE = "image"
-
-
-class OpenRewardPromptBlock(_StrictModel):
-    """Serializable prompt block snapshot from an OpenReward session."""
-
-    type: OpenRewardPromptBlockType
-    text: str | None = None
-    data: str | None = None
-    mime_type: str | None = None
-    detail: JSONDict | None = None
-
-    @model_validator(mode="after")
-    def _validate_block_shape(self) -> "OpenRewardPromptBlock":
-        if self.type == OpenRewardPromptBlockType.TEXT:
-            if self.text is None:
-                raise ValueError("text blocks must include text")
-            if self.data is not None or self.mime_type is not None:
-                raise ValueError("text blocks cannot include image payload fields")
-            return self
-
-        if self.data is None or self.mime_type is None:
-            raise ValueError("image blocks must include data and mime_type")
-        if self.text is not None:
-            raise ValueError("image blocks cannot include text")
-        return self
-
-
-class OpenRewardToolSpec(_StrictModel):
-    """Serializable snapshot of a tool exposed by an OpenReward session."""
-
-    name: str
-    description: str
-    input_schema: JSONDict | None = None
-
-
-class OpenRewardTaskManifestEntry(_StrictModel):
-    """A single task snapshot with the prompt and tools seen by the agent."""
-
-    task_index: int
-    task_spec: JSONDict
-    prompt_blocks: list[OpenRewardPromptBlock]
-    tools: list[OpenRewardToolSpec]
-
-
-class OpenRewardTaskManifest(_StrictModel):
-    """Snapshot of a deterministic task subset for one OpenReward split."""
-
-    deployment_name: str
-    environment_name: str
-    split: str
-    tasks: list[OpenRewardTaskManifestEntry]
-
-    @property
-    def task_count(self) -> int:
-        return len(self.tasks)
+from marin.rl.integrations.openreward.client import load_openreward_client
+from marin.rl.integrations.openreward.models import (
+    OpenRewardPromptBlock,
+    OpenRewardPromptBlockType,
+    OpenRewardTaskManifest,
+    OpenRewardTaskManifestEntry,
+    OpenRewardToolSpec,
+    SecretsMapping,
+)
 
 
 class OpenRewardTaskLike(Protocol):
@@ -275,16 +212,6 @@ def load_openreward_task_manifest(input_path: str) -> OpenRewardTaskManifest:
 
     with open_url(input_path, "rb") as fd:
         return OpenRewardTaskManifest.model_validate_json(fd.read())
-
-
-def load_openreward_client() -> type:
-    try:
-        from openreward import OpenReward
-    except ImportError as exc:
-        raise ImportError(
-            "OpenReward SDK is not installed. Install `openreward` before preparing task manifests."
-        ) from exc
-    return OpenReward
 
 
 def prepare_openreward_task_manifest(
